refactor RHI

+++ sources/libs/renderer/tavros/renderer/internal/opengl/command_list_opengl.hpp
@@ -26,9 +26,9 @@ namespace tavros::renderer
 
         void draw_indexed(uint32 index_count, uint32 first_index = 0, uint32 vertex_offset = 0, uint32 instance_count = 1, uint32 first_instance = 0) override;
 
-        void copy_buffer_data(buffer_handle buffer, const void* data, uint64 size, uint64 offset = 0) override;
+        void copy_buffer_data(buffer_handle buffer, const void* data, size_t size, size_t offset = 0) override;
 
-        void copy_buffer(buffer_handle dst_buffer, buffer_handle src_buffer, uint64 size, uint64 dst_offset = 0, uint64 src_offset = 0) override;
+        void copy_buffer(buffer_handle dst_buffer, buffer_handle src_buffer, size_t size, size_t dst_offset = 0, size_t src_offset = 0) override;
 
     private:
         graphics_device_opengl* m_device = nullptr;
